✨ Implement source for lexemes
